Adds a new tab-content.scss document containing all of the inner tab styles. These replace several removed styles from the tabs.scss document.
Simplifies the templates for tab content on the main character sheet and updates the classes to use the new styles.

Updates the tab helper to use the new classes.

Removes the biography specific styles which have been replaced with a --bio modifier.

Adds a new utility class for setting an element's height to 100%.