"Send extra to chat" feature.
Added same button to "extra", I forgot about it.
Changed the class  and function names. Now they are not so specific about "stunts"
While I was there I explicitly set async:"false" in all .enrichHTML calls if it is not set to "true". In foundry 10 we got warning and in foundry 11 default behaivor was reversed,  and we got [object Object] instead of text. This should fix the problem.